[aergo.system] add address member to the voting power data
There exists two kinds of IDs in Aergo. They are address and account ID. The
former is used to send tokens by token owners, while the latter is inervally
used as a key for the state DB.

Before this commit, a voting power struct contain only an account ID. Now it
also includes Aergo address. Although an account ID is only needed for giving
voting reward, the corresponding address is added to voting power data for
debugging purpose.