[consensus] handle timeout tx
If a TX fails to execute by timeout, it may constantly fail afterwards. To
prevent this, a such TX is marked on its first error and included as the
first TX of the next block. If it fails again, it will be eviected.

This commit only includes the first half of the implementation (mark and
reschedule).