Ensure aergoluac has a non-zero exit code on error
Always return non-zero exit code on error for consistency.  This
ensures that if aergoluac is called from within shell scripts /
Makefiles etc., errors can be properly detected and handled.

Returning the error from the cobra.Command closure automatically
causes it to be output to STDERR, so we don't need to print it
ourselves.