Update batch.go
Returning a return code is a good idea.
However, I modified the code a bit, because when you reset the `batchErrorCount` at the beginning, the error count was initialized when another batch file is called from the module called batch, so I removed all the code that initialized the `batchErrorCount`.
And at the end of the print, error is already printed like 'Batch is failed: Error <err number>'. Right now It is confused by printing an error every time the batch level changes.